Fix several compiler warnings
	-Fix make jobserver complaints
	-Fix a lzma param shadowing warning (many more remain)
	-Fix comma no effect warnings
	-Fix comparison warning
	-Fix unused expression warning on assert++
	-Make num_blocks assertion conditional on NDEBUG
	-More fix warnings

+++ lib/sparsewrite.cc
@@ -365,7 +365,7 @@ bool SparseWrite(int fd,
              /*fprintf(stderr, "But writing %04llX..%04llX from %p\n", \
                  BufferedPos, BufferedPos+BufferedSize-1, BufferedBegin); */ \
             ssize_t res = pwrite64(fd, BufferedBegin, BufferedSize, BufferedPos); \
-            if(res != BufferedSize) return false; \
+            if((size_t)res != BufferedSize) return false; \
         } \
         BufferedSize = 0; BufferedBegin = Buffer; BufferedPos = WritePos; \
     } while(0)
